+++ packages/crypto/tsup.config.ts
@@ -1,4 +1,4 @@
 import { defineConfig } from 'tsup';
 import base from '../../tsup.config';
 
-export default defineConfig(base);
+export default defineConfig({ ...base, entry: ['src/**/*.{ts,tsx}'], bundle: false });