+++ packages/zstd/src/zstd.ts
@@ -1,4 +1,5 @@
-import { Buffer } from 'buffer';
+import { nodeZstd } from './node-zstd';
+import { rnZstd } from './react-native-zstd';
 
 interface ZstdModule {
   compress(data: Uint8Array): Uint8Array;
@@ -18,26 +19,8 @@ export class Zstd {
 }
 
 function getAvailableZstdModule(): ZstdModule {
-  // 1. Try to use Node.js 'zstd-napi' package
-  try {
-    const nodeZstd = require('zstd-napi');
-    if (nodeZstd) return nodeZstd;
-  } catch {
-    // Ignore
-  }
-
-  // 2. Try to use 'react-native-zstd' package (for React Native environments)
-  try {
-    const rnZstd = require('react-native-zstd');
-    if (rnZstd)
-      return {
-        // Additional data conversion is needed due to inconsistent API between Node and RN implementations
-        compress: (data: Uint8Array): Uint8Array => rnZstd.compress(Buffer.from(data).toString()),
-        decompress: (data: Uint8Array): Uint8Array => Uint8Array.from(Buffer.from(rnZstd.decompress([...data]))),
-      };
-  } catch {
-    // Ignore
-  }
+  if (nodeZstd) return nodeZstd;
+  else if (rnZstd) return rnZstd;
 
   throw new Error(
     "No available zstd module found. Please install 'zstd-napi' or 'react-native-zstd' depending on a platform"