+++ packages/cache/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hiero-did-sdk/cache
 
+## 0.1.5
+
+### Patch Changes
+
+- Use CJS for React Native + split optional dependencies loading
+- Updated dependencies
+  - @hiero-did-sdk/core@0.1.5
+
 ## 0.1.4
 
 ### Patch Changes